# apps/backend-api/swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Login realizado com sucesso."},"401":{"description":"Credenciais inválidas."}},"summary":"Realiza login do usuário","tags":["auth"]}},"/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"Usuário criado com sucesso."},"400":{"description":"Dados inválidos."}},"summary":"Cria um novo usuário","tags":["users"]},"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":"Lista de usuários retornada com sucesso."}},"summary":"Lista todos os usuários","tags":["users"]}},"/users/active":{"get":{"operationId":"UsersController_findActiveUsers","parameters":[],"responses":{"200":{"description":"Lista de usuários ativos retornada com sucesso."}},"summary":"Lista usuários ativos","tags":["users"]}},"/users/search":{"get":{"operationId":"UsersController_searchByName","parameters":[{"name":"name","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Usuários encontrados."},"404":{"description":"Nenhum usuário encontrado."}},"summary":"Busca usuários pelo nome","tags":["users"]}},"/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Usuário encontrado."},"404":{"description":"Usuário não encontrado."}},"summary":"Busca usuário por ID","tags":["users"]},"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"Usuário atualizado com sucesso."},"404":{"description":"Usuário não encontrado."}},"summary":"Atualiza dados do usuário","tags":["users"]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Usuário removido com sucesso."},"404":{"description":"Usuário não encontrado."}},"summary":"Remove usuário","tags":["users"]}},"/users/{id}/activate":{"patch":{"operationId":"UsersController_activate","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Usuário ativado com sucesso."},"404":{"description":"Usuário não encontrado."}},"summary":"Ativa usuário","tags":["users"]}},"/users/{id}/deactivate":{"patch":{"operationId":"UsersController_deactivate","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Usuário desativado com sucesso."},"404":{"description":"Usuário não encontrado."}},"summary":"Desativa usuário","tags":["users"]}}},"info":{"title":"A initial sytem API","description":"The Initial system API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"joao@email.com","description":"E-mail do usuário"},"password":{"type":"string","example":"Senha123","description":"Senha do usuário"}},"required":["email","password"]},"CreateUserDto":{"type":"object","properties":{"name":{"type":"string","example":"João Silva","description":"Nome completo do usuário"},"email":{"type":"string","example":"joao@email.com","description":"E-mail do usuário"},"password":{"type":"string","example":"Senha123","description":"Senha do usuário. Deve ter pelo menos 8 caracteres, uma letra maiúscula, uma minúscula e um número."},"isActive":{"type":"boolean","example":true,"description":"Indica se o usuário está ativo"}},"required":["name","email","password"]},"UpdateUserDto":{"type":"object","properties":{}}}}}